Fix the collision system
